Move flavor text to a dedicated util file
